uint to uint256

+++ hardhat/contracts/amm_pool/CurveProxy.sol
@@ -126,10 +126,10 @@ contract CurveProxy is Initializable, RelayRecipient {
     }
 
     struct SwapExactTokensParams {
-        uint amountOutMin;
+        uint256 amountOutMin;
         address path;
         address to;
-        uint deadline;
+        uint256 deadline;
     }
 
     struct MetaTokenParams {
